All elements now resize properly
